requirements.txt
cryptography
Django
djangorestframework
djangorestframework-jwt
psycopg2-binary
gunicorn
dynaconf
